chore: theme toggle switch generalized
