+++ src/components/Button/index.tsx
@@ -1,0 +1,5 @@
+import { ButtonContainer } from './styles'
+
+export function Button() {
+  return <ButtonContainer />
+}
