+++ src/components/Button/styles.ts
@@ -1,0 +1,3 @@
+import styled from 'styled-components'
+
+export const ButtonContainer = styled.button``
