+++ src/components/Footer/index.tsx
@@ -1,7 +1,8 @@
-import { ThemeToggleSwitch } from '../ToggleSwitch'
+import { Moon, Sun } from 'phosphor-react'
 
 import { FooterContainer } from './styles'
 import { useThemeSwitch } from '../../contexts'
+import { IconToggleSwitch } from '../ToggleSwitch'
 
 export function Footer() {
   const { isDefaultTheme, changeTheme } = useThemeSwitch()
@@ -10,7 +11,11 @@ export function Footer() {
     <FooterContainer>
       <span>Em desenvolvimento por Bernardo Costa Nascimento.</span>
 
-      <ThemeToggleSwitch toggled={isDefaultTheme} onToggle={changeTheme} />
+      <IconToggleSwitch
+        toggled={isDefaultTheme}
+        onToggle={changeTheme}
+        icons={[Moon, Sun]}
+      />
     </FooterContainer>
   )
 }
