+++ src/components/ToggleSwitch/interfaces/IconToggleSwitchProps.ts
@@ -1,0 +1,7 @@
+import { Icon } from 'phosphor-react'
+
+import ToggleSwitchProps from './ToggleSwitchProps'
+
+export default interface IconToggleSwitchProps extends ToggleSwitchProps {
+  icons: [Icon, Icon]
+}
