+++ src/components/ToggleSwitch/interfaces/ToggleSwitchProps.ts
@@ -1,0 +1,4 @@
+export default interface ToggleSwitchProps {
+  toggled: boolean
+  onToggle: () => void
+}
